requirements.txt
Pillow>=10.0.0
pywin32>=306
pystray>=0.19.5
pyinstaller>=6.0.0
pytest>=8.0.0
